Added new models and simple parser

+++ src/main/java/io/github/pabulaner/jxsd/input/value/ElementValue.java
@@ -2,13 +2,15 @@ package io.github.pabulaner.jxsd.input.value;
 
 import io.github.pabulaner.jxsd.input.type.ValueType;
 
+import java.math.BigInteger;
+
 public class ElementValue extends Value {
 
     private final ValueType type;
 
     private final String name;
 
-    public ElementValue(int minOccurs, int maxOccurs, ValueType type, String name) {
+    public ElementValue(BigInteger minOccurs, BigInteger maxOccurs, ValueType type, String name) {
         super(Kind.SINGLE, minOccurs, maxOccurs);
 
         this.type = type;
